# apps/api/storage/framework/testing-commerce-openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Watch",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "security": [
        {
            "sessionCookie": []
        }
    ],
    "paths": {
...
        "/api/v1/cart/items": {
            "post": {
                "operationId": "cart.storeItem",
                "tags": [
                    "Cart"
                ],
                "parameters": [
                    {
                        "name": "X-Cart-Token",
                        "in": "header",
                        "required": true,
                        "description": "Opaque guest cart token.",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCartItemRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Updated guest cart.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CartResource"
                                        },
                                        "meta": {
                                            "type": "object",
                                            "properties": {
                                                "cart_token": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "cart_token"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                },
                "security": []
            }
        },
        "/api/v1/cart/items/{itemId}": {
            "patch": {
                "operationId": "cart.updateItem",
                "tags": [
                    "Cart"
                ],
                "parameters": [
                    {
                        "name": "itemId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-Cart-Token",
                        "in": "header",
                        "required": true,
                        "description": "Opaque guest cart token.",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCartItemRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated guest cart.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CartResource"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                },
                "security": []
            },
            "delete": {
                "operationId": "cart.destroyItem",
                "tags": [
                    "Cart"
                ],
                "parameters": [
                    {
                        "name": "itemId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-Cart-Token",
                        "in": "header",
                        "required": true,
                        "description": "Opaque guest cart token.",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Removed guest cart item.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "null"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
        "/api/v1/customer/cart": {
            "get": {
                "operationId": "cart.customerShow",
                "tags": [
                    "Cart"
                ],
                "responses": {
                    "200": {
                        "description": "Customer cart.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CartResource"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
        "/api/v1/customer/cart/items": {
            "post": {
                "operationId": "cart.customerStoreItem",
                "tags": [
                    "Cart"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCartItemRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Updated customer cart.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CartResource"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
        "/api/v1/customer/cart/items/{itemId}": {
            "patch": {
                "operationId": "cart.customerUpdateItem",
                "tags": [
                    "Cart"
                ],
                "parameters": [
                    {
                        "name": "itemId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCartItemRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated customer cart.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CartResource"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "cart.customerDestroyItem",
                "tags": [
                    "Cart"
                ],
                "parameters": [
                    {
                        "name": "itemId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Removed customer cart item.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "null"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "csrfHeader": {
                "type": "apiKey",
                "description": "CSRF token from the XSRF-TOKEN cookie (Sanctum SPA).",
                "in": "header",
                "name": "X-XSRF-TOKEN"
            },
            "sessionCookie": {
                "type": "apiKey",
                "description": "Sanctum SPA session cookie.",
                "in": "cookie",
                "name": "laravel_session"
            }
        },
        "schemas": {
...
            "CartItemResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "product_variant_id": {
                        "type": "integer"
                    },
                    "qty": {
                        "type": "integer"
                    },
                    "unit_price": {
                        "type": "string"
                    },
                    "line_total": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "product_variant_id",
                    "qty",
                    "unit_price",
                    "line_total"
                ],
                "title": "CartItemResource"
            },
            "CartResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "currency": {
                        "type": "string"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CartItemResource"
                        }
                    },
                    "subtotal": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "currency",
                    "items",
                    "subtotal"
                ],
                "title": "CartResource"
            },
...
            "StoreCartItemRequest": {
                "type": "object",
                "properties": {
                    "product_variant_id": {
                        "type": "integer"
                    },
                    "qty": {
                        "type": "integer",
                        "minimum": 1
                    }
                },
                "required": [
                    "product_variant_id",
                    "qty"
                ],
                "title": "StoreCartItemRequest"
            },
...
            "UpdateCartItemRequest": {
                "type": "object",
                "properties": {
                    "qty": {
                        "type": "integer",
                        "minimum": 1
                    }
                },
                "required": [
                    "qty"
                ],
                "title": "UpdateCartItemRequest"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}